--- tests/store/test_store_cli.py
@@ -1,1 +1,0 @@
-# -*- coding: utf-8 -*-
